Fixed errors if no channels were added to vote

+++ vote/vote.py
@@ -54,29 +54,31 @@ class vote(commands.Cog):
     @listener()
     async def on_message(self, message):
         channels = await self.config.guild(message.guild).voting()
-        if message.channel.id in channels:
-            await message.add_reaction(self.bot.get_emoji(await self.config.channel(message.channel).upemote()))
-            await message.add_reaction(self.bot.get_emoji(await self.config.channel(message.channel).downemote()))
+        if channels:
+            if message.channel.id in channels:
+                await message.add_reaction(self.bot.get_emoji(await self.config.channel(message.channel).upemote()))
+                await message.add_reaction(self.bot.get_emoji(await self.config.channel(message.channel).downemote()))
 
     @listener()
     async def on_reaction_add(self, reaction, user):
         channels = await self.config.guild(reaction.message.guild).voting()
-        if reaction.message.channel.id in channels:
-            up = await self.config.channel(reaction.message.channel).upemote()
-            down = await self.config.channel(reaction.message.channel).downemote()
-            pin = await self.config.channel(reaction.message.channel).pin()
-            remove = await self.config.channel(reaction.message.channel).remove()
-            if reaction.emoji.id == up or reaction.emoji.id == down:
-                if user == reaction.message.author:
-                    await reaction.remove(user)
-                if reaction.emoji.id == up and pin != 0:
-                    if reaction.count > pin:
-                        print("pinned a great meme")
-                        await reaction.message.pin()
-                if reaction.emoji.id == down and remove != 0:
-                    if reaction.count > remove:
-                        print("deleted a bad meme")
-                        await reaction.message.delete()
+        if channels:
+            if reaction.message.channel.id in channels:
+                up = await self.config.channel(reaction.message.channel).upemote()
+                down = await self.config.channel(reaction.message.channel).downemote()
+                pin = await self.config.channel(reaction.message.channel).pin()
+                remove = await self.config.channel(reaction.message.channel).remove()
+                if reaction.emoji.id == up or reaction.emoji.id == down:
+                    if user == reaction.message.author:
+                        await reaction.remove(user)
+                    if reaction.emoji.id == up and pin != 0:
+                        if reaction.count > pin:
+                            print("pinned a great meme")
+                            await reaction.message.pin()
+                    if reaction.emoji.id == down and remove != 0:
+                        if reaction.count > remove:
+                            print("deleted a bad meme")
+                            await reaction.message.delete()
